+++ atl-checker/src/atl/game_formula.rs
@@ -2,8 +2,8 @@ use std::fmt::{Display, Formatter};
 
 use joinery::prelude::*;
 
-use crate::atl::formula::Phi;
-use crate::atl::gamestructure::GameStructure;
+use crate::atl::Phi;
+use crate::game_structure::GameStructure;
 
 /// The GamePhi struct pairs an ATL formula with its game structure, allowing us to print
 /// the formula using the names of players and labels that is defined by the game structure.
