+++ atl-checker/src/game_structure/lcgs/ir/intermediate.rs
@@ -1,19 +1,17 @@
 use std::borrow::BorrowMut;
 use std::collections::{HashMap, HashSet};
 
-use crate::atl::common;
-use crate::atl::common::{Action, Proposition};
-use crate::atl::formula::{identifier, ATLExpressionParser};
-use crate::atl::gamestructure::GameStructure;
-use crate::lcgs::ast::{ConstDecl, Decl, DeclKind, ExprKind, Identifier, Root};
-use crate::lcgs::ir::error::Error;
-use crate::lcgs::ir::eval::Evaluator;
-use crate::lcgs::ir::relabeling::Relabeler;
-use crate::lcgs::ir::symbol_checker::{CheckMode, SymbolChecker, SymbolError};
-use crate::lcgs::ir::symbol_table::{Owner, SymbolIdentifier, SymbolTable};
+use crate::atl::{identifier, ATLExpressionParser};
+use crate::game_structure;
+use crate::game_structure::lcgs::ast::{ConstDecl, Decl, DeclKind, ExprKind, Identifier, Root};
+use crate::game_structure::lcgs::ir::error::Error;
+use crate::game_structure::lcgs::ir::eval::Evaluator;
+use crate::game_structure::lcgs::ir::relabeling::Relabeler;
+use crate::game_structure::lcgs::ir::symbol_checker::{CheckMode, SymbolChecker, SymbolError};
+use crate::game_structure::lcgs::ir::symbol_table::{Owner, SymbolIdentifier, SymbolTable};
+use crate::game_structure::{Action, GameStructure, Proposition};
 use pom::parser::{sym, Parser};
 use std::fmt::{Display, Formatter};
-use std::sync::Arc;
 
 /// A struct that holds information about players for the intermediate representation
 /// of the lazy game structure
@@ -142,7 +140,7 @@ impl IntermediateLCGS {
     pub(crate) fn available_actions(
         &self,
         state: &State,
-        player: common::Player,
+        player: game_structure::Player,
     ) -> Vec<SymbolIdentifier> {
         self.players[player]
             .actions
@@ -442,7 +440,7 @@ impl GameStructure for IntermediateLCGS {
     }
 
     /// Returns the set of labels/propositions available in the given state.
-    fn labels(&self, state: common::State) -> HashSet<Proposition> {
+    fn labels(&self, state: game_structure::State) -> HashSet<Proposition> {
         let state = self.state_from_index(state);
         let mut res = HashSet::new();
 
@@ -462,7 +460,7 @@ impl GameStructure for IntermediateLCGS {
     }
 
     /// Returns the next state given a current state and an action for each player.
-    fn transitions(&self, state: common::State, choices: Vec<usize>) -> usize {
+    fn transitions(&self, state: game_structure::State, choices: Vec<usize>) -> usize {
         let mut state = self.state_from_index(state);
         // To evaluate the next state we assign the actions to either 1 or 0 depending
         // on whether or not the action was taken
@@ -504,7 +502,7 @@ impl GameStructure for IntermediateLCGS {
     }
 
     /// Returns the number of moves available to each player in the given state.
-    fn move_count(&self, state: common::State) -> Vec<usize> {
+    fn move_count(&self, state: game_structure::State) -> Vec<usize> {
         let state = self.state_from_index(state);
         self.players
             .iter()
@@ -513,7 +511,7 @@ impl GameStructure for IntermediateLCGS {
             .collect()
     }
 
-    fn state_name(&self, state: common::State) -> String {
+    fn state_name(&self, state: game_structure::State) -> String {
         self.state_from_index(state).to_string()
     }
 
@@ -521,11 +519,16 @@ impl GameStructure for IntermediateLCGS {
         self.labels.get(proposition).unwrap().to_string()
     }
 
-    fn player_name(&self, player: common::Player) -> String {
+    fn player_name(&self, player: game_structure::Player) -> String {
         self.players.get(player).unwrap().name.to_string()
     }
 
-    fn action_name(&self, state: common::State, player: common::Player, action: Action) -> String {
+    fn action_name(
+        &self,
+        state: game_structure::State,
+        player: game_structure::Player,
+        action: Action,
+    ) -> String {
         let state = self.state_from_index(state);
         let actions = self.available_actions(&state, player);
         actions.get(action).unwrap().to_string()
@@ -533,7 +536,7 @@ impl GameStructure for IntermediateLCGS {
 }
 
 impl ATLExpressionParser for IntermediateLCGS {
-    fn player_parser(&self) -> Parser<u8, common::Player> {
+    fn player_parser(&self) -> Parser<u8, game_structure::Player> {
         // In ATL, players are referred to using their name, i.e. an identifier
         identifier().convert(move |name| {
             // We check if a declaration with the given name exists,
@@ -592,12 +595,12 @@ impl ATLExpressionParser for IntermediateLCGS {
 
 #[cfg(test)]
 mod test {
-    use crate::atl::gamestructure::GameStructure;
-    use crate::lcgs::ast::DeclKind;
-    use crate::lcgs::ir::intermediate::{IntermediateLCGS, State};
-    use crate::lcgs::ir::symbol_table::Owner;
-    use crate::lcgs::ir::symbol_table::SymbolIdentifier;
-    use crate::lcgs::parse::parse_lcgs;
+    use crate::game_structure::lcgs::ast::DeclKind;
+    use crate::game_structure::lcgs::ir::intermediate::{IntermediateLCGS, State};
+    use crate::game_structure::lcgs::ir::symbol_table::Owner;
+    use crate::game_structure::lcgs::ir::symbol_table::SymbolIdentifier;
+    use crate::game_structure::lcgs::parse::parse_lcgs;
+    use crate::game_structure::GameStructure;
     use std::collections::HashMap;
 
     #[test]
@@ -607,13 +610,10 @@ mod test {
         const max_health = 100;
         player alice = gamer;
         player bob = gamer;
-
         template gamer
             health : [0 .. max_health] init max_health;
             health' = health - 1;
-
             label alive = health > 0;
-
             [wait] 1;
             [shoot] health > 0;
         endtemplate
@@ -660,13 +660,10 @@ mod test {
         const max_health = 100;
         player anna = gamer [enemy=bob];
         player bob = gamer [enemy=anna];
-
         template gamer
             health : [0 .. max_health] init max_health;
             health' = enemy.shoot ? health - 1 : health;
-
             label alive = health > 0;
-
             [wait] 1;
             [shoot] health > 0 && enemy.health > 0;
         endtemplate
@@ -693,7 +690,6 @@ mod test {
         let input = "
         player anna = human [var=apples, act=dance, prop=happy];
         player bob = human [var=bananas, act=run, prop=sad];
-
         template human
             var : [0 .. 10] init 5;
             var' = var;
@@ -720,7 +716,6 @@ mod test {
         let input = "
         player anna = human [act=work, income=1000 + work * 200, expenses=1000];
         player bob = human [income=1000, expenses=money * 2 / 10];
-
         template human
             money : [0 .. 10000] init 2000;
             money' = money + income - expenses;
@@ -805,11 +800,10 @@ mod test {
         foo' = foo;
         bar : [-2 .. 5] init 3;
         bar' = bar;
-        
+
         player p1 = test;
         player p2 = test;
         player p3 = test;
-
         template test
             yum : [0 .. 8] init 0;
             yum' = yum;
@@ -1005,13 +999,12 @@ mod test {
         // Can we update state even though it depends on unavailable actions?
         let input = "
         player ryan = guy;
-        
+
         some_var : [0 .. 10] init 0;
         some_var' = some_var + ryan.unavailable_action;
-
         template guy
             [unavailable_action] 0;
-            [available_action] 1; 
+            [available_action] 1;
         endtemplate
         ";
         let lcgs = IntermediateLCGS::create(parse_lcgs(input).unwrap()).unwrap();
