+++ atl-checker/src/game_structure/lcgs/ir/mod.rs
