+++ atl-checker/src/game_structure/mod.rs
@@ -1,11 +1,16 @@
 use std::collections::HashSet;
 
-#[allow(unused_imports)]
 pub use eager::EagerGameStructure;
 
-use crate::atl::common::{Action, Player, Proposition, State};
+use std::sync::Arc;
 
 mod eager;
+pub mod lcgs;
+
+pub(crate) type Proposition = usize;
+pub(crate) type Player = usize;
+pub(crate) type State = usize;
+pub(crate) type Action = usize;
 
 pub trait GameStructure {
     fn max_player(&self) -> usize;
@@ -28,4 +33,30 @@ pub trait GameStructure {
 
     /// Returns the human-readable name of the given action by the given player in the given state
     fn action_name(&self, state: State, player: Player, action: Action) -> String;
+}
+
+/// Implements Vec of Vecs in arbitrary runtime determined depth.
+#[derive(Clone, Debug, Deserialize)]
+#[serde(untagged)]
+pub enum DynVec {
+    NEST(Vec<Arc<DynVec>>),
+    BASE(State),
+}
+
+/// Indexes into a DynVec.
+/// The length of `choices` must match the depth of `transitions`.
+pub(crate) fn transition_lookup(choices: &[usize], transitions: &DynVec) -> State {
+    match transitions {
+        DynVec::NEST(v) => {
+            if choices.is_empty() {
+                panic!("Fewer choices given than number of players in transitions");
+            }
+
+            let choice = choices[0];
+            let h: &DynVec = v.get(choice).expect("Out of bounds choice");
+
+            transition_lookup(&choices[1..choices.len()], h)
+        }
+        DynVec::BASE(state) => *state,
+    }
 }
