+++ atl-checker/src/lib.rs
@@ -7,13 +7,12 @@ extern crate tracing;
 
 #[macro_use]
 mod simple_edg;
+#[macro_use]
+mod emit_count;
+pub mod algorithms;
 pub mod analyse;
 pub mod atl;
-mod com;
-pub mod common;
 pub mod edg;
-pub mod lcgs;
+pub mod game_structure;
 #[cfg(feature = "graph-printer")]
 pub mod printer;
-pub mod search_strategy;
-pub mod solve_set;
